Title: 최솟값 3개, Time: 182ms, Memory: 1MB, Status: Passed - CodeTree

+++ 240809/최솟값 3개/top-3-smallest-number.cpp
@@ -12,14 +12,14 @@ int main() {
         int num;
         cin >> num;
         pq.push(-num);
-        
+
         if(pq.size()<3) cout << -1 <<'\n';
         else{
-            int a = -pq.top();
+            long long a = -pq.top();
             pq.pop();
-            int b = -pq.top();
+            long long b = -pq.top();
             pq.pop();
-            int c = -pq.top();
+            long long c = -pq.top();
             pq.pop();
 
             cout << a*b*c <<'\n';
